requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
asgiref==3.7.2
    # via django
colorama==0.4.6
    # via pytest
django==4.2.3
    # via
    #   -r requirements.in
    #   djangorestframework
djangorestframework==3.14.0
    # via -r requirements.in
exceptiongroup==1.1.2
    # via pytest
iniconfig==2.0.0
    # via pytest
packaging==23.1
    # via pytest
pluggy==1.2.0
    # via pytest
pytest==7.4.0
    # via
    #   -r requirements.in
    #   pytest-django
pytest-django==4.5.2
    # via -r requirements.in
pytz==2023.3
    # via djangorestframework
sqlparse==0.4.4
    # via django
tomli==2.0.1
    # via pytest
typing-extensions==4.7.1
    # via asgiref
tzdata==2023.3
    # via django
